removed board.rb file.. move board methods to the main.rb file
